Add syslog support to alarm-notify.sh.
This uses the logger command (found on all modern Linux and BSD systems)
to allow alarm-notify.sh to send netdata alarm messages to a local
or remote syslog server.  By default, logging is done to the local6
facility at a warning log level, with critical alarms being logged at
the err level.

This provides a mechanism of easily generating a running log of all
netdata alarm events on most standard UNIX systems, and when used with
full-featured logging daemons like rsyslog or syslong-ng can be used to
archive such messages in a wide varity of alternative formats.

There are a few limitations currently:

 * It can only log to one location.  This can be easily worked around in
 most syslog daemon's configurations, as almost all of them support some
 form of output multiplexing.
 * The default settings will usually not get the results most people
 would want, but because of the sometimes very significant differences
 in default syslog configurations from distribution to distribution,
 there's not really any way to make the defaults any saner.
 * While it is possible to log directly to a remote syslog server, there
 is no quick and easy wet up for this, it requires the user to manually
 compose the extra options that need to be passed to the logger command.

Syslog messages generated by this will look something like this:

    Apr  3 09:00:30 localhost netdata: Netdata notification on wild-karde at Tue Apr 3 09:00:00 EDT 2018: WARNING, out of disk space time = 5h