update openrc init script
Quote variables where able.
Switch to a more posix double 'and'.
Remove unused variables.
The exit timeout was fixed.

We also remove fetching the configuration from the init script
(it should not live in it).

Further we define reload rotate and save commands to prevent the need to
restart all the time.