Update syslog support to allow per-role configuration.
This adds support to syslog notifications to have per-role
configuration.  This includes specifying facilty and log-level per-role,
target servers per-role, and a message prefix per-role, possibly
multiple times.  This is done by specifying each combination of
configuration as a 'recipient' in the role recipients variables using
the following format:

[[facility.level][@host[:port]]/]prefix

As part of this, the default log level has been changed to depend solely
on the type of alert, crit for CRITICAL, warning for WARNING, and info
for everything else (most notably CLEAR), and the end-user configuration
and the user configuration of these default levels has been removed.

In addition, the message format has been changed to look like the
following:

    Apr  3 09:00:30 localhost netdata: prefix WARNING on wild-karde at Tue Apr 3 09:00:00 EDT 2018: out of disk space time = 5h