Add Amazon SNS notification support.
The Simple Notification Service (SNS) is a reasonably simple message
broker service provided by Amazon as part of it's AWS offerings.

SNS utilizes the concept of 'topics' (similar to Netdata's concept of
'roles' for notifications) to control message routing.  Any given topic
may have any number of subscribers of any of the following types:

* Email addresses.
* Phone numbers for SMS.
* HTTP or HTTPS web hooks.
* AWS Lambda endpoints.
* AWS SQS endpoints.
* Mobile applications (via various native push notification services).

Topics are rpersented by Amazon Resource Names (ARN), which are a special
type of URI.

Unfortunately, properly signing requests to AWS endpoints is a serious
pain in the arse, so we pretty much have to use the CLI interface.
Because of the inflexibility of this tool, setup is somewhat painful.

THis uses topic ARN's directly as recipients.  SNS does not support
delivery to multiple topics in bulk, so a separate call is required for
each topic ARN.  Users are provided with the ability to customize the
message format used for the notifications.