ExecutableService: retrun [] instead of None if no data. (#4346)
This fixes a bug in the logind module causing it to throw errors when
there are no users logged in.

No net change for any other in-tree ExecutableService users.  All of
them either use truth value testing or depend on errors being thrown
when trying to subscript the return value, both of which behave the same
for an empty list and None.