netdata/packaging/installer: HoS situation - Fix broken install-or-update script (#5806)
During the last major installer refactoring, we managed to break (again) the static64 installer.
We modified the code that defines user/group ownerships for netdata installation,
forcing the install to end up with a process running as netdata and files set as root

To fix this, we re-instate default NETDATA_USER/NETDATA_GROUP to root, as it should be.
Then we attempt group netadata creation. If that succeeds, we attempt user netdata creation.
If that succeeds, we attempt secondary groups addition.

Report errors on all otherwise situations from each step.