Add alarm status counter api call (#6554)
##### Summary
This is implementation of a prerequisite for the requested feature #6536 (Generate an overall status badge/chart for the health of category)

##### Component Name
web/api/
health/

##### Details
Provide a new, `alarm_count` API call that returns the total number of alarms for given contexts and alarm states. Default is the total number of raised alarms, for all contexts.