Inject archived backports repository on Debian Jessie for CI package builds. (#7495)
* Inject archived backports repository on Debian Jessie.

Debian Jessie only ever had libuv1 as a backport, not part of the base
release. This means that we need the backports repository on Jessie for
the build to work, which has been archived.

This updates the environment prep script for building Debian packages
for Jessie to inject a file into `/etc/apt/sources.list.d` to add in the
archived repository so that everything works correctly.

* Change how the archived backports are injected for jessie.

This shifts to just appending to the main sources.list file, which
should be more reliable. It also explicitly updates caches after
changing the sources.

* Switch back to using a separate sources file.

* Moved Jessie archived backports handling to repo setup function.

This makes the code a bit cleaner overall.

Also, we need to disable validity checking for the archives during the
call to `apt-get update` otherwise things will just refuse to work.

* Fix syntax error.