Improve system-info.sh to better handle certain cases when gathering info on the system's disk capacity. (#7902)
* Remove trailing whitespace in system-info.sh.

* Fix handling of APFS on macOS.

APFS can have multiple volumes in a single partition, which means that
the same functional 'volume' can appear multiple times in the output of
`df`. Duplicate lines for such volumes will show the same total size and
available space along with a common prefix for the device name.

This updates the parsing logic for `df` on macOS to account for this by
deduplicating lines in the `df` output that have the same total size,
available space, and same normalized device name.

This has the potential to incorrectly under-account space in some cases,
but the liklihood of that happeing is much less than the certainty of
overaccounting space on standard APFS configurations.

* Properly handle VirtIO block devices when using /sys.

The VirtIO Block device driver uses a dynamically allocated device major
number, meaning that we can't trivially match on it.

This updates the handling to properly look it up in `/proc/devices`
instead of just using the whole dynamic device number range.

* Add handling for NVMe block devices in sysfs code.

They use dynamic major numbers just like VirtIO Block devices do.

* Switch to device major discovery in /proc/devices for all device types.

This converts the code to use `/proc/devices` to look up correct device
major numbers for block devices taht we treat as disks just like we are
already doing for those that have dynamically assigned numbers. This
makes the code both more robust and easier to understand and modify.

This also excludes some particularly old hardware that we were
originally looking for. If needed, we can add in the required device
names, but for now it's better to keep the list concise.

* Correct handling of device major discovery.

We need to strip leading whitespace before calling cut, not after.

* Only use /sys/block if we can read /proc/devices.

We use `/proc/devices` to do device number lookups that we then use to
filter devices under `/sys/block`. As a result, if we can't read
`/proc/devices`, then we won't actually parse anything out of
`/sys/block` either, so we need to just fall back to parsing `df`
output.

* Deduplicate `df` output by device name on Linux.

This ensures that we properly handle BTRFS subvolumes, counting each
actual volume only once.

* Use POSIX math expansion instead of awk to sum disk sizes.

This avoids the rather annoying habit of AWK of printing integers in
scientific notation instead of as exact values.

* Correct `sed` options for POSIX complaince.

* Fix disk info fetching for macOS>

POSIX tools, as found on macOS, lack a number of rather useful filtering
and sorting features, so we need to get rather creative with the
handling on macOS to make the disk space computation work correctly.

This unfortunately makes the calculation a bit less reliable than it
would have been had the existing calculations worked correctly, but it's
the best I can come up with without making things exponentiall more
complicated.

* Properly handle sector size when using sysfs.