Fixed condition controlling use of static LWS in RPM builds. (#10661)
The original intent was to only use a static copy of LWS if we couldn’t
use a system copy. Due to two incorrect conditionals in the spec file
though, what was actually happening was that we were using a static
build of LWS on systems where we could use a system copy, and not
linking against LWS at all on other systems.

This fixes those conditionals so that we use a system copy if available,
and provide a static copy if a system copy is not available.