Assorted cleanup in the OpenRC init script. (#13115)
* Properly handle service dependencies in OpenRC init script.

This removes the `NETDATA_START_AFTER_SERVICES` variable from our OpenRC
script. This variable made the script somewhat harder to read, and did
not actually provide any additional functionality (users can simply
define appropriate `rc_after` values in `/etc/conf.d/netdata` to achieve
exactly the same net effect)..

* Provide more concrete descriptions for OpenRC commands.

Updates the command descriptions in the OpenRC init script to properly
describe exactly what they are doing, instead of going with generic and
rather ambiguous names.

* Allow use of netdatacli for additional commands in OpenRC init script.

This makes life simpler for users who want to run the agent under
OpenRC’s native process supervision, which does not allow use of
`start-stop-daemon` commands in the init script.

Existing behavior (using `start-stop-daemon`) is preserved as the
default.

* Fix background command handling.

THe `pidfile` variable should only be set if the agent is not being run
under `supervise-daemon`.

* COnsolidate code from additional commands, and handle supervise-daemon correctly.

* Fix typo.

* Remove pointless `require_files` line.

We don’t actually do anything in the script that cares about the config
file existing, and the agent starts fine without it, so we should not be
requiring it to be present.