enable libunwind in static builds (#19764)
* enable libunwind in static builds

* add libunwind and backtrace to buildinfo

* add linunwind to alpine packages

* add -dev packages

* add remove libunwind binary from the packages

* Vendor libunwind in static builds instead of using a copy from the build environment.

This is required to ensure that the C++ exception handling functionality
in libunwind is _disabled_, because it does not play nice with static
linking when using C++ with exception handling support enabled.

* Remove changes from local testing.

* Fix cross architecture builds.

* Disable libunwind on 64-bit POWER builds.

musl libc does not include functions that are required to build
libunwind for this platform, so just disable it there for now.

---------

Co-authored-by: Austin S. Hemmelgarn <austin@netdata.cloud>