Integrate OpenTelemetry collector build into build system. (#19702)
* Fix Go version requirement detection to not have external deps.

Instead of relying on a UNIX-like environment with the `grep` and `cut`
commands, perform the required data extraction using CMake code. This
makes it more portable, but also should speed things up a tiny bit
because it doesn’t require invoking external commands.

* Preliminary integration of the new otel-collector into the build.

Adding `-DENABLE_PLUGIN_OTEL=On` to CMake options will enable building
the plugin and installing it.

Currently disabled by default, and does not include packaging
integration yet.

The plugin itself can be built independently of the primary build system
but using the same build infrastructure that is used to build it as
part of the regular build by using CMake in the src/go/otel-collector
directory.

* Minor cleanup.

* Fix build prefix.

* Restructure to not use a sub-project.