Switch to using a relative RUNPATH for IBM plugin library lookup. (#21131)
We can’t rely on CMAKE_INSTALL_PREFIX at build time because it may be
different at install time from what was specified at build time.

Instead, set the RUNPATH relative to the binary itself. This is done
using the `$ORIGIN` keyword in the RUNPATH to reference the directory
the plugin is located in and then using `..` to reference parent
directories.