Report failures to update native packages. (#21144)
* Report failures to update native packages.

The current check is relatively naive and only compares the installed
version of Netdata against the latest published release. Long term we
should probably improve it, but for now this is good enough to clearly
report the inability to update.

* Only flag cases where installed version is unchanged.

This avoids the most likely case of false positives (updating while
we’re in the process of publishing a new version).

* Address review comments.

* Assorted fixes.

* Use 4 digits for major version in updater checks.