Use systemd-tmpfiles to handle required directories at runtime when possible. (#21243)
* Use systemd-tmpfiles to handle required directories.

This pushes handling of permissions and ownership of our required
directories to systemd-tmpfiles when possible, which helps persistently
ensure that they are correct, and allows us to make at least some Netdata
service files both simpler and more reliable.

* Address review comments.

* Fix typos.

* Update directory creation to be more robust.

* Update tmpfiles config to more robustly ensure correct permissions.

This adds new entries for the cloud.d and registry directories to be
created, as well as adding entries to apply correct permissions and
ownership for database files, claiming configuration, and a handful of
other files.