Report inability to update due to native packages no longer being published. (#21288)
* Report inability to update due to native packages no longer being published.

This adds an extra check in the `netdata-updater.sh` script to fail with
an error when trying to update a system that uses our native packages if
we have stopped publishing our native packages.

The check utilizes a flag file generated by the repository backend code
to determine if we are actually publishing packages for the platform or
not. If the flag file is not present, then the update process bails very
early on with an error message indicating that the user needs to change
to a different install type, and a link pointing to our documentation of
how to switch install types.

* Address review comments.

* Skip version change check if no release channel can be determined.

Fixes: #21328

* Fix incorrect variable name.

* Add missing SYSVERSION definition.

* Use correct version flag for version checks in updater.

* Don’t duplicate fetching of installed agent version in updater.

In the native package handling code, instead of invoking `netdata -V` a
second time, just use the parsed version number to determine if
something is a nightly build or a stable build.

This matches on whether the final five digits of the prased version
number are all zeroes (which indicates a stable build) or not.